RVT_parse: e-mail treatment has been improved for PST/OST (through RVT_sanitize_libpff_item and RVT_sanitize_libpff_attachment) and for other mail types (through RVT_parse_eml).
An additional .RVT_metadata is NO LONGER created for each e-mail item. This information (message metadata / headers) is contained within the HTML itself; it is initially hidden when opening the HTML file, but clicking the link in the first line or pressing any letter displays this section (needs JavaScript).

This results in a lower number of files (half, indeed) being created in the filesystem. Moreover, as a consequence of being inside the HTML and not in a separate .RVT_metadata file, this content is now indexed and searched on. For instance, host or domain names appearing only in the SMTP headers of a message will now show up in a search.

In the case of RVT_parse_eml, the RVT_metadata section (needed for tracing item source when searching over the text atoms) has been placed inside the message HTML representation, starting on line 2. RVT_get_source has been modified to acknowledge this. For PFF (PST/OST) items, this is not needed as that information lies on output/parser/control/pff-XX.RVT_metadata.

The HTML code generated in RVT_parse_eml (affects at least EML, DBX, mbox, and MSG extensions), now include CSS code as in RVT_sanitize_libpff_item (PST/OST items).

The HTML code generated for e-mails now shows the attachment size (in bytes) next to the HREF link to each attachment.